refactor: moved mock server to v3 so deps can be installed during versioned test run
