+++ tests/versioned/v3/ai-server/responses/amazon.js
