+++ tests/versioned/v3/ai-server/responses/claude.js
