+++ tests/versioned/v3/ai-server/testdata/stream.expected.bin
